refactor: Create a support module for vm tests
